0530_AgenticAIforStartupswithAzure/session-1-azure-toolbox/demo/requirements.txt
# Core Azure SDK packages
azure-search-documents==11.5.1
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-storage-blob==12.21.0

# OpenAI integration
openai==1.40.6

# Document processing
pypdf2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.3
requests==2.32.3

# Web framework for demo
fastapi==0.112.0
uvicorn==0.30.5
jinja2==3.1.4
python-multipart==0.0.9

# Utilities
python-dotenv==1.0.1
pydantic==2.8.2
